Move docker base into new directory
