+++ buildstockbatch/cloud/docker_base.py
